addition added

+++ Calculator.java
@@ -6,8 +6,8 @@ class Calculator{
 	public static void main(String[] agrs){
 		Scanner scanner = new Scanner(System.in);
 		System.out.println("Enetr nuber for addition");
-		double a = scanner.nextDouble();
-		double b = scanner.nextDouble();
+		int a = scanner.nextInt();
+		int b = scanner.nextInt();
 		
 		System.out.println(add(a,b));
 }}
